# Backend/api/documentation/openapi.yaml
openapi: 3.0.3
info:
  title: 'Kebab Api'
  description: ''
  version: 1.0.0
servers:
  -
    url: api/documentation
paths:
  /api/register:
    post:
      summary: ''
      operationId: postApiRegister
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: ''
                  example: est
                  nullable: false
                email:
                  type: string
                  description: 'Must be a valid email address.'
                  example: igibson@example.com
                  nullable: false
                password:
                  type: string
                  description: ''
                  example: corporis
                  nullable: false
              required:
                - name
                - email
                - password
      security: []
  /api/login:
    post:
      summary: ''
      operationId: postApiLogin
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Must be a valid email address. The <code>email</code> of an existing record in the users table.'
                  example: kirlin.tara@example.com
                  nullable: false
                password:
                  type: string
                  description: ''
                  example: expedita
                  nullable: false
              required:
                - email
                - password
      security: []
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []